MOR-452: Moved the top part of the editor into a component
