+++ src/views/MorphicBarEditor.vue
@@ -36,141 +36,18 @@
         <div id="barEditor" class="pt-2">
 
           <!-- Bar info, actions, and editor tabs -->
-          <div id="EditorTop" class="border-bottom">
-            <!-- details and tabs -->
-            <div>
-              <div id="BarDetails">
-                <!-- Bar name -->
-                <div class="bar-name">
-                  <h2>
-                    {{barName}}
-
-                    <!-- rename bar -->
-                    <span v-if="barDetails.name !== 'Default'">
-                      <TextInputDialog id="barNameDialog"
-                                       title="Rename Bar"
-                                       prompt="Enter the new name for the bar"
-                                       v-model="barDetails.name"
-                                       @ok="renameBar"
-                      />
-                      &nbsp;<small><b-button variant="link" v-b-modal="'barNameDialog'">rename</b-button></small>
-                    </span>
-                  </h2>
-
-                </div>
-                <div class="lead">
-                  <template v-if="barMembers.length === 0">
-                    Shared bar
-                  </template>
-                  <template v-else-if="barMembers.length === 1">
-                    Bar for: <b>{{ barMembers[0].displayName }}</b>
-                  </template>
-                  <template v-else>
-                    Bar for: <b>{{ barMembers.length }} members</b>
-                  </template>
-                </div>
-              </div>
-
-              <div id="EditorTabs">
-                <b-tabs class="editorTabs"
-                        v-model="editorTabIndex"
-                        small
-                        :content-class="'bg-white border border-top-0 ' + (editorTabIndex ? '' : 'd-none')">
-
-                  <!-- hidden tab to simulate no tab being selected -->
-                  <b-tab title="" active title-item-class="d-none" class="d-none" />
-
-                  <!-- Members tab -->
-                  <b-tab @click="editorTabIndex = (editorTabIndex === 1 ? 0 : 1)">
-                    <template #title>
-                      <b-icon-person-circle/>&nbsp;
-                      <span v-if="activeMemberId">{{ memberDetails.displayName }} details ({{memberDetails.stateText}})</span>
-                      <span v-else-if="memberCount === 0">Unused Bar</span>
-                      <span v-else>Members ({{ memberCount }})</span>
-                    </template>
-                    <button @click="editorTabIndex = 0" type="button" aria-label="Close" class="close">×</button>
-
-                    <b-card>
-                      <b-card-title>
-                        <span v-if="activeMemberId"><b-icon-person-circle /> {{ memberDetails.first_name }}</span>
-                        <span v-else-if="memberCount === 1"><b-icon-person-circle /> This Morphic Bar is used by 1 member</span>
-                        <span v-else-if="memberCount > 0"><b-icon-people-fill /> This Morphic Bar is used by {{memberCount}} members</span>
-                        <span v-else>This Morphic Bar is NOT used</span>
-                      </b-card-title>
-
-                      <!-- member's own bar -->
-                      <ul v-if="activeMemberId" class="list-unstyled">
-                        <li v-if="memberDetails.role === 'member'"><b-button variant="link" v-b-modal.roleChangeConfirm>Make member a Group Manager</b-button></li>
-                        <li v-else><b-button variant="link" v-b-modal.roleChangeConfirm>Remove group manager role from member</b-button></li>
-                        <li><b-button variant="link" v-b-modal.deleteConfirm class="text-danger">Delete member</b-button></li>
-
-                        <li v-if="!memberDetails.isCurrent"><b-button variant="link" v-b-modal="'inviteMemberDialog'">{{ memberDetails.state === 'uninvited' ? "Send Invitation" : "Re-invite member" }}</b-button></li>
-                      </ul>
-
-                      <!-- community bar -->
-                      <ul v-else-if="memberCount > 0" >
-                        <li v-for="member in members" v-bind:key="member.id">
-                          {{ member.first_name }} {{ member.last_name }}
-                        </li>
-                      </ul>
-
-                      <!-- unused community bar -->
-                      <b-card-sub-title v-else>You can go back to the <b-link to="/dashboard">Dashboard</b-link> and invite members to use it.</b-card-sub-title>
-                    </b-card>
-
-                  </b-tab>
-
-                  <!-- Bar settings tab -->
-                  <b-tab @click="editorTabIndex = (editorTabIndex === 2 ? 0 : 2)" disabled title-item-class="notProduction">
-                    <template #title>
-                      <b-icon-gear-fill/> Morphic Bar settings
-                    </template>
-                    <button @click="editorTabIndex = 0" type="button" aria-label="Close" class="close">×</button>
-
-                    <b-card>
-                      <b-card-title><b-icon-gear-fill/> Morphic Bar settings</b-card-title>
-
-                      <b-form-checkbox id="barOnRight" v-model="bar.settings.barOnRight" name="barOnRight" value="true" unchecked-value="false">
-                        Bar on the right of the screen
-                      </b-form-checkbox>
-                      <b-form-checkbox id="cannotClose" v-model="bar.settings.cannotClose" name="cannotClose" value="true" unchecked-value="false">
-                        Member cannot close bar
-                      </b-form-checkbox>
-                      <b-form-checkbox id="startsOpen" v-model="bar.settings.startsOpen" name="startsOpen" value="true" unchecked-value="false">
-                        Morphic Bar always starts open
-                      </b-form-checkbox>
-                    </b-card>
-                  </b-tab>
-
-                </b-tabs>
-
-              </div>
-            </div>
-
-            <div id="EditorActions">
-              <b-button variant="outline-dark"
-                        v-b-modal="'copyBarDialog'"
-                >Copy bar from...</b-button>
-
-              <b-button variant="outline-dark"
-                        :disabled="!isChanged || newBar"
-                        @click="revertBar"
-                >Revert to user's current bar</b-button>
-
-              <b-button variant="success"
-                        style="visibility: hidden"
-                        disabled
-                >Try it</b-button>
-
-              <b-button variant="success"
-                        :disabled="!isChanged"
-                        @click="saveBar"
-                >Save bar</b-button>
-            </div>
-          </div>
+          <EditorDetails ref="EditorDetails"
+                         :bar-details="barDetails"
+                         :bars-list="barsList"
+                         :bar-members="barMembers"
+                         :is-changed="isChanged"
+                         :member-details="memberDetails"
+                         :new-bar="newBar"
+                         @save-bar="saveBar()" @revert-bar="revertBar()"
+                         />
 
           <!-- the desktop -->
-          <div id="preview-holder" class="desktop mt-3" @click="editorTabIndex = 0">
+          <div id="preview-holder" class="desktop mt-3" @click="$refs.EditorDetails.closeTab()">
             <drop mode="cut" class="dragToDelete desktop-portion">
               <template v-slot:drag-image="">
                 <img src="/img/trash.svg" style="height: 100px; width: 100px; margin-left: -50px; margin-top: -50px"/>
@@ -291,99 +168,6 @@
   #barEditor {
     padding-left: 15px;
     padding-right: 15px;
-
-    #EditorTop {
-      display: flex;
-      flex-direction: row;
-      flex-wrap: nowrap;
-
-      & > :first-child {
-        flex-grow: 1;
-      }
-    }
-
-    #BarDetails {
-      flex-grow: 1;
-    }
-
-    #EditorActions {
-      display: grid;
-      align-content: center;
-      grid-template-columns: auto auto;
-
-      column-gap: 0.5rem;
-      row-gap: 0.5rem;
-
-      .btn {
-        font-size: 0.9rem;
-        min-width: max-content;
-      }
-
-    }
-  }
-
-  #EditorTabs {
-    display: flex;
-    align-items: center;
-
-    & > :not(:last-child) {
-      margin-right: 1em;
-    }
-
-    .editorTabs {
-      margin-bottom: -1px;
-      .tab-content {
-        position: absolute;
-        z-index: 10;
-        max-width: 70%;
-        min-width: 25rem;
-        border-radius: 0 3px 3px 3px;
-
-        & > div {
-          margin-top: 0;
-          padding: 0.3rem;
-        }
-
-      }
-      .hidden-tab {
-        display: none;
-      }
-      .card {
-        border: 0;
-        .card-body {
-          padding: 0.5rem;
-        }
-      }
-
-    }
-
-    .barSelection {
-      width: unset !important;
-    }
-
-    .userInvitationStatusArea {
-      flex-grow: 1;
-      text-align: right;
-
-      .dot {
-        height: 13px;
-        width: 13px;
-        display: inline-block;
-        border-radius: 50%;
-        vertical-align: middle;
-
-        &.active-state { background-color: #3bc03b; }
-        &.invited-state { background-color: #f3c702; }
-        &.uninvited-state { background-color: #f20202; }
-      }
-
-      .linkStyling {
-        border: none;
-        background: inherit;
-        color: #069;
-        cursor: pointer;
-      }
-    }
   }
 
   #preview-holder.desktop {
@@ -651,18 +435,18 @@ import { Drag, Drop, DropList } from "vue-easy-dnd";
 import * as Bar from "@/utils/bar";
 import EditButtonDialog from "@/components/editor/EditButtonDialog";
 import BarItemLink from "@/components/editor/BarItemLink";
-import TextInputDialog from "@/components/dialogs/TextInputDialog";
 import ButtonCatalog from "@/components/editor/ButtonCatalog";
 import InviteMemberDialog from "@/components/dialogs/InviteMemberDialog";
 import CopyBarDialog from "@/components/dialogs/CopyBarDialog";
+import EditorDetails from "@/components/editor/EditorDetails";
 
 export default {
     name: "MorphicBarEditor",
     components: {
+        EditorDetails,
         ButtonCatalog,
         CopyBarDialog,
         InviteMemberDialog,
-        TextInputDialog,
         BarItemLink,
         EditButtonDialog,
         SidePanel,
@@ -1018,17 +802,6 @@ export default {
                             this.barsList = barsData;
                             this.membersList = resp.data.members;
                             this.membersList = this.membersList.map(m => { return m.bar_id ? m : Object.assign(m, { bar_id: this.community.default_bar_id }); });
-                            this.members = [];
-
-                            if (resp.data.members.length > 0) {
-                                for (let i = 0; i < resp.data.members.length; i++) {
-                                    if (this.$route.query.barId === resp.data.members[i].bar_id) {
-                                        this.members.push(resp.data.members[i]);
-                                    }
-                                }
-                            }
-
-                            this.memberCount = this.members.length;
                         })
                         .catch(err => {
                             console.error(err);
@@ -1037,12 +810,6 @@ export default {
                 .catch(err => {
                     console.error(err);
                 });
-        },
-        getBarRemoveValidity: function () {
-            if (this.barDetails.name !== "Default" && this.memberCount === 0) {
-                return true;
-            }
-            return false;
         },
         getCommunityData: function () {
             getCommunity(this.communityId)
@@ -1066,26 +833,7 @@ export default {
                 list.push(bar);
             });
         },
-        /**
-         * OK button on the bar name dialog clicked.
-         * @param {TextInputOKEvent} event The event object.
-         */
-        renameBar: function (event) {
-            // only saving the name - so, get the currently stored bar and change that.
-            event.promise = getCommunityBar(this.communityId, this.barDetails.id).then(resp => {
-                /** @type {BarDetails} */
-                const bar = resp.data;
-                bar.name = event.newValue;
-                return updateCommunityBar(this.communityId, this.barDetails.id, bar).then(() => {
-                    // Update the item in the full list, to reload it in the community manager component.
-                    const listed = this.barsList.find(b => b.id === this.barDetails.id);
-                    if (listed) {
-                        listed.name = bar.name;
-                    }
-                    return true;
-                });
-            });
-        },
+
         /**
          * Gets the edit route for a bar.
          * @param {BarDetails} bar The bar.
@@ -1169,13 +917,7 @@ export default {
             return this.activeMemberId && this.membersList.find(m => m.id === this.activeMemberId);
         },
 
-        editDialog: function () { return this.$refs.editDialog; },
-        /**
-         * @return {String} The bar title.
-         */
-        barName: function () {
-            return Bar.getBarName(this.barDetails);
-        }
+        editDialog: function () { return this.$refs.editDialog; }
 
     },
     mounted() {
@@ -1227,7 +969,6 @@ export default {
             this.storeUnsavedBar();
         },
         "$route.query": function () {
-            this.members = [];
             this.initialChangesPrimaryItems = false;
             this.initialChangesDrawerItems = false;
             this.loadAllData();
@@ -1256,7 +997,6 @@ export default {
             openDrawer: true,
             editDialogShown: false,
             tab: 0,
-            editorTabIndex: 0,
             dragFromEditor: false,
             isChanged: false,
             editBarName: false,
@@ -1279,7 +1019,8 @@ export default {
              * @type {Array<BarDetails>}
              */
             userBars: [],
-            /** @type {Array<CommunityMember>} */
+            /** All community members
+             * @type {Array<CommunityMember>} */
             membersList: [],
 
             /** @type {ButtonCatalog} Button catalog. */
@@ -1306,9 +1047,6 @@ export default {
             barDetails: {},
             originalBarDetails: {},
 
-            memberCount: 0,
-            members: [],
-
             // configurations
             preview: {
                 drawer: {
@@ -1323,13 +1061,6 @@ export default {
                 name: "New Bar",
                 is_shared: false,
                 items: []
-            },
-            bar: {
-                settings: {
-                    barOnRight: true,
-                    cannotClose: false,
-                    startsOpen: false
-                }
             },
             predefinedBars: predefinedBars,
             checkBarTimer: null
